fix for freezing monitor, inheriting basic command and reorganized code.
Found the initial commit I added was from an archived version of the command. This should better reflect the current command we are using.

Some edits to fix redundant parts in the message handler and made the command base from BasicCommand instead of CLxCommand.

Also exposing more arguments that was requested from artists and hopefully living up to the eula requirement making it clear this is using UV Packmaster