Ignore hidden polygons
Fixed many "No Valid Static Island" errors where users wanted to pack a subset of islands while having deselected islands outside of packing area.

Did away with checks for user having any polygons selected and instead set a bool if a polygon is not selected and visible.

Users will still get the error when attempting to "mark" an island as static that is outside of the packing area.